Some maintenance done (removed all warnings, some references for primitive types etc) and vertices of non-visible blocks are no longer added to the terrain mesh.

+++ MineCloneProject/src/MCP/ImGuiLayer.cpp
@@ -131,7 +131,7 @@ namespace MC
 	bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent& e)
 	{
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2((float)e.GetWidth(), (float)e.GetHeight());
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		RenderCommand::SetViewport(0, 0, e.GetWidth(), e.GetHeight());
